Replace queryblock and queryinput
