+++ src/components/energiatodistus-haku/querybuilder/queryblock.svelte
@@ -1,4 +1,5 @@
 <script>
+  import qs from 'qs';
   import * as R from 'ramda';
   import * as Maybe from '@Utility/maybe-utils';
   import Radio from '@Component/Radio/Radio';
@@ -8,83 +9,31 @@
   import * as Inputs from '@Component/Energiatodistus/inputs';
   import * as EtHakuUtils from '@Component/energiatodistus-haku/energiatodistus-haku-utils';
 
-  import NumberOperatorInput from '@Component/energiatodistus-haku/querybuilder/query-inputs/number-operator-input';
-  import OperatorInput from '@Component/energiatodistus-haku/querybuilder/query-inputs/operator-input';
-  import BooleanInput from '@Component/energiatodistus-haku/querybuilder/query-inputs/boolean-input';
-  import DateInput from '@Component/energiatodistus-haku/querybuilder/query-inputs/date-input';
+  import QueryInput from './query-input';
 
-  import QueryInput from '@Component/energiatodistus-haku/querybuilder/query-input';
-
-  export let model;
-  export let lens;
-
+  export let operator;
+  export let key;
+  export let values;
+  export let index;
   export let schema;
-
-  const mapOperatorTypeToComponent = type =>
-    R.prop(type, {
-      [EtHakuUtils.OPERATOR_TYPES.STRING]: OperatorInput,
-      [EtHakuUtils.OPERATOR_TYPES.NUMBER]: NumberOperatorInput,
-      [EtHakuUtils.OPERATOR_TYPES.DATESINGLE]: DateInput,
-      [EtHakuUtils.OPERATOR_TYPES.BOOLEAN]: BooleanInput
-    });
-
-  const conjunctionLens = R.compose(
-    lens,
-    R.lensProp('conjunction')
-  );
-
-  const blockLens = R.compose(
-    lens,
-    R.lensProp('block')
-  );
-
-  const operationLens = R.compose(
-    blockLens,
-    R.lensIndex(0)
-  );
-
-  const keyLens = R.compose(
-    blockLens,
-    R.lensIndex(1)
-  );
-
-  let conjunction = R.view(conjunctionLens, model);
-
-  let key = R.view(keyLens, model);
-
-  $: model = R.set(conjunctionLens, conjunction, model);
-
-  $: operations = R.prop(key, schema);
 </script>
 
-<div class="w-full flex flex-col">
-  {#if Maybe.isSome(conjunction)}
-    <div class="flex justify-between w-1/6 my-10">
-      <Radio
-        bind:group={conjunction}
-        value={Maybe.Some('and')}
-        label={$_('energiatodistus.haku.and')}
-        on:click={evt => (conjunction = Maybe.Some('and'))} />
-      <Radio
-        bind:group={conjunction}
-        value={Maybe.Some('or')}
-        label={$_('energiatodistus.haku.or')}
-        on:click={evt => (conjunction = Maybe.Some('or'))} />
-    </div>
-  {/if}
-
-  <div class="flex-grow flex items-center justify-start">
-    <div class="w-1/2 mr-4">
-      <Select
-        items={R.keys(schema)}
-        bind:model={key}
-        format={Inputs.propertyLabel($_)}
-        parse={R.identity}
-        lens={R.identity}
-        allowNone={false} />
-    </div>
-    <div class="w-1/2">
-      <QueryInput bind:model lens={blockLens} {operations} />
-    </div>
+<div class="flex-grow flex items-center justify-start">
+  <div class="w-1/2 mr-4">
+    <Select
+      name={`${index}_key`}
+      items={R.keys(schema)}
+      bind:model={key}
+      format={Inputs.propertyLabel($_)}
+      parse={R.identity}
+      lens={R.identity}
+      allowNone={false} />
+  </div>
+  <div class="w-1/2">
+    <QueryInput
+      nameprefix={`${index}`}
+      operation={operator}
+      {values}
+      operations={R.prop(key, schema)} />
   </div>
 </div>
